extract request errors into a module
